fix(ingest/data-lake): address review feedback on DuckDB profiling
- Restore an actionable install hint when DuckDB is missing (the s3-slim extra
  omits the data-lake-profiling deps) instead of a bare ModuleNotFoundError that
  aborts the run.
- Move DataLakeProfilerConfig into the shared data_lake_common layer (where the
  DuckDB profiler lives) and re-export it from s3 and abs. The abs profiling
  config no longer diverges from its siblings, and the profiler no longer
  type-imports config from the s3 connector.
- Use constant warning messages + stable titles + context for the per-table
  profiling warnings, so the run-summary dedup key no longer fragments per
  path/limit/dataset URN.
- Drop the unused DuckDB get_quantiles_expr override (dead code:
  get_column_quantiles builds the expression directly).
- Remove gcs from the profiling SupportedSources lists: GCS profiling is not
  configurable (no profiling field, capability marked unsupported).
- Clear the adapter's per-table SUMMARIZE cache at setup so a reused adapter
  instance cannot serve another table's stats.
